Allow user to specify search radius

+++ src/main/java/com/davewhoyt/bg/data/model/Location.java
@@ -31,8 +31,8 @@ public class Location {
     private Long locationId;
 
     @Column private String name;
-    @Column private BigDecimal latitude;
-    @Column private BigDecimal longitude;
+    @Column private Double latitude;
+    @Column private Double longitude;
 
     /** calculated by some queries*/
     @Transient
@@ -70,22 +70,22 @@ public class Location {
     }
 
 
-    public BigDecimal getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
 
-    public void setLatitude(BigDecimal latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
 
-    public BigDecimal getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
 
-    public void setLongitude(BigDecimal longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
